refactor base64decode and fix panic
